requirements.txt
feedparser==6.0.10
gspread==5.12.4
oauth2client==4.1.3
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
Pillow==10.0.1
google-api-python-client==2.108.0
google-auth==2.23.4
lxml==4.9.3
urllib3==2.0.7